% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayola.R
\name{scale_fill_crayola}
\alias{scale_fill_crayola}
\title{Crayola colour}
\usage{
scale_fill_crayola(...)
}
\arguments{
\item{...}{Additional parameters to pass to  \code{scale_fill_manual}.}
}
\description{
Crayola colour
}

