% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_crayola.R
\name{scale_colour_crayola}
\alias{scale_colour_crayola}
\title{Crayola colour}
\usage{
scale_colour_crayola(n = 100, ...)
}
\arguments{
\item{n}{Number of repeats of the colours, default set to 100}

\item{...}{Additional parameters to pass to  \code{scale_fill_manual}.}
}
\description{
Crayola colour
}
