% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_crayola.R
\name{scale_fill_crayola}
\alias{scale_fill_crayola}
\title{Crayola colour}
\usage{
scale_fill_crayola(n = 100, guide = "none")
}
\arguments{
\item{n}{Number of repeats of the colours, default set to 100}

\item{guide}{guide (default "none", rather than "legend")}
}
\description{
Crayola colour
}
